Error

+++ disk-raspios-enable-ssh.sh
@@ -32,8 +32,12 @@ if [ -z "$1" ] || [ "$1" = "-h" ] || [ "$1" = "--help" ]; then
 fi
 TMP_MNT_PATH="tmp/raspios"
 sudo diskutil unmountDisk "$1"
-mkdir -p "$TMP_MNT_PATH"
-sudo mount "$1" "$TMP_MNT_PATH"
-touch "$TMP_MNT_PATH/ssh"
-sudo diskutil unmountDisk "$1"
-rm -rf "$TMP_MNT_PATH"
+if mkdir "$TMP_MNT_PATH"; then
+    sudo mount "$1" "$TMP_MNT_PATH"
+    touch "$TMP_MNT_PATH/ssh"
+    sudo diskutil unmountDisk "$1"
+    rm -rf "$TMP_MNT_PATH"
+else
+    echo "-- Error while creating $TMP_MNT_PATH"
+    exit 1
+fi
